Implemented backup enhancement

+++ proj1/src/workers/DeleteFileWorker.java
@@ -4,13 +4,18 @@ import messages.DeleteMessage;
 import peer.Peer;
 import utils.Utils;
 
+import java.util.concurrent.ScheduledThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
+
 public class DeleteFileWorker implements Runnable {
     private final Peer peer;
     private final String fileId;
+    private final ScheduledThreadPoolExecutor scheduler;
 
     public DeleteFileWorker(Peer peer, String fileId) {
         this.peer = peer;
         this.fileId = fileId;
+        this.scheduler = new ScheduledThreadPoolExecutor(1);
     }
 
     @Override
@@ -19,14 +24,21 @@ public class DeleteFileWorker implements Runnable {
 
         // Try to send DELETE message max 3 times
         int attempt = 0;
-        do {
-            this.peer.sendControlMessage(deleteMessage);
-            //System.out.printf("Sent DELETE for file %s\n", this.fileId);
-
-            int wait = (int) Math.pow(2, attempt) * 1000;
-            Utils.sleep(wait);
-        } while (++attempt < Utils.MAX_3_ATTEMPTS);
-
-        this.peer.getStorage().deleteSentChunks(this.fileId);
+        this.scheduler.schedule(() -> this.sendDeleteMessage(deleteMessage, attempt), (int) Math.pow(2, attempt) * 1000, TimeUnit.MILLISECONDS);
     }
+
+    private void sendDeleteMessage(DeleteMessage deleteMessage, int attempt){
+        this.peer.sendControlMessage(deleteMessage);
+
+        if(attempt<Utils.MAX_3_ATTEMPTS) {
+            int finalAttempt = ++attempt;
+            this.scheduler.schedule(() -> this.sendDeleteMessage(deleteMessage, finalAttempt), (int) Math.pow(2, attempt) * 1000, TimeUnit.MILLISECONDS);
+        }
+        else{
+            this.peer.getStorage().deleteSentChunks(this.fileId);
+        }
+
+    }
+
+
 }
